Moved check log in to App.js so that you dont have to check log in on every page. Same with Navbar and username.
