+++ src/Home/Question.js
@@ -8,7 +8,7 @@ import {
 import './CoursePage.css';
 import axios from 'axios';
 import * as firebase from 'firebase';
-import Navbar from "./Navbar";
+// import Navbar from "./Navbar";
 // import Calendar from "./Calendar2";
 // import Calendar from 'rc-calendar';
 // import LuxonUtils from '@date-io/luxon';
@@ -29,8 +29,8 @@ class Question extends PureComponent {
         }
         this.checkLoggedIn = this.checkLoggedIn.bind(this);
         this.handleChange = this.handleChange.bind(this);
-        this.logout = this.logout.bind(this);
-        this.getUserName = this.getUserName.bind(this);
+        // this.logout = this.logout.bind(this);
+        // this.getUserName = this.getUserName.bind(this);
         this.getQuestion = this.getQuestion.bind(this);
         this.submitAnswer = this.submitAnswer.bind(this);
     }
@@ -49,7 +49,7 @@ class Question extends PureComponent {
                 this.getQuestion,
                 this.getReplies(questionID)
             )
-        this.getUserName();
+        // this.getUserName();
         
     }
 
@@ -68,25 +68,25 @@ class Question extends PureComponent {
         });
     };
     
-    logout(e){
-        e.preventDefault();
-        firebase.auth().signOut();
-        this.props.history.push('/');
-    }
-    getUserName(e){
-        var page = this;
-        firebase.auth().onAuthStateChanged(function(user) {
-        if (user) {
-            var info = {
-                id: user.uid
-            }
-            axios.post('https://studybuddys-223920.appspot.com/api/getUsername', info)
-            .then(response => {
-                page.setState({name: response.data.name})
-            })
-        }
-        });
-    }
+    // logout(e){
+    //     e.preventDefault();
+    //     firebase.auth().signOut();
+    //     this.props.history.push('/');
+    // }
+    // getUserName(e){
+    //     var page = this;
+    //     firebase.auth().onAuthStateChanged(function(user) {
+    //     if (user) {
+    //         var info = {
+    //             id: user.uid
+    //         }
+    //         axios.post('https://studybuddys-223920.appspot.com/api/getUsername', info)
+    //         .then(response => {
+    //             page.setState({name: response.data.name})
+    //         })
+    //     }
+    //     });
+    // }
 
     submitAnswer() {
         var reply = this.state.replyText;
@@ -136,13 +136,13 @@ class Question extends PureComponent {
     render() {
         return (
             <div data-aos ="fade-in" data-aos-easing="linear" data-aos-duration="800" style = {{display: "flex", flexDirection: "column"}}>
-                <div>
+                {/* <div>
                     <div style = {{float: "right", display: "inline-block"}}>
                         <span>{this.state.name}</span>
                         <Button onClick={this.logout}>Logout</Button>
                     </div>
                     <Navbar/>
-                </div>
+                </div> */}
                     {/* <Button 
                         className="Calendar"
                         type="submit"
